added spinner next to progressBar
spinner added next to the progressBar. The code for the spinner can be found in progressBar.ts, with two simple functions to start and stop the spinner.

The progressBar width was slightly reduced, and the background color was changed to have better contrast and better show the progress (it is easier to tell what 100% of the progessBar is this way).
This was needed because with the addition of the spinner it was hard to tell where the progressBar ends.