Added detailed errors option in settings, implemented showing detailed erros vs non-detailed errors (whole line long)
Added a detailed error option in settings, if enabled detailed errors are shown (errors are shown at the exact characters they occur), if disabled errors are shown over the whole line where the error occurs.
The actual implementation (at Editor.ts) uses two functions which look at the raw file and determine the offset needed to reach the start and end of the line (offsetToStartOfLine and offsetToEndOfLine at Editor.ts:427 and Editor.ts:435 respectively). If detailed errors are disabled, this offset gets added to the start and end points of the error.
For this to work, a new private variable _rawContents was created which stores the raw file contents.