Client: Use the beginning of the sentence when executing commands. (#38)
* Client: Use the beginning of the sentence when executing commands.

The beginning of the current sentence (ie. the sentence that the
cursor is currently in) refers to the end position of the previous
sentence, since we do not have enough information to actually find
the beginning.

* SentenceManager: Handle edge case when cursor at end of line.

Handles the case where the cursor is at the end of the line, i.e.
after the `.` character. The position after the `.` still belongs to
the line so we increment the line by one.