Added coq-syntax option to change tactics which show up for autocompletion (#40)
* Added coq-syntax option to change tactics which show up

Added new tacticsCoq.json file with the normal coq-syntax tactics.

Also added a setting to change coq-syntax. The information for the selected syntax is passed to editor.ts, which initializes the tactics.ts singleton accordingly.

Laslty, implemented singleton method in tactics.ts to use normal or coqTactics based on how the singleton is initialized by editor.ts
Editor.ts "knows" what to initialize the singleton to based on the setting selected by the user.
Moreover, the singleton can be "overriden" (i.e., can be initialized more than once) such that if the setting changes, the singleton will change accordingly.