Version Checker: Do not start the extension without language server.
The version checker checks to see if `waterproof.path` points at a
valid `coq-lsp` binary, if not then we will not start the extension.
An error message is shown to the user and the statusbar item turns red,
indicating that the Waterproof extension has stopped.

Changing the `Waterproof.path` setting and clicking the statusbar item
allows the user to restart the extension and try again without having
to restart vscode.