added check for periods in expand definition
Whenever the given input ends with a period followed by 0 or more spaces and/or new lines, we disregard them.
This only happens if the sentence ends with that combination, so a period in the middle of the sentence is not removed.
This also fixes the bug where if you add new lines on the second input box extra lines and characters will appear in the file you are editing.