separated commonExecute functionality into three buttons for Help, Search and Expand Definition
Separated the commonExecute functionality, covering Help, Search and Expand Definition into three, also separating the button into three different buttons.

Added fix for extra empty lines and/or a dot at the end of the command for Search.

Additionally, fixed error with the Expand Definition panel breaking down if no document was "linked" to it.