New ProseMirror document constructor (#73)
* Parse input and hint blocks

* Add support for mathdisplay and coq top level nodes

* Add support for top level markdown

* Add support for coqdoc

* Turn blocks into prosemirror nodes

* Add tests for inner block extraction

* Add basic support for `.v` files

* Experimental: Construct document using new parser

* Menubar: Split cases of '.v' vs '.mv' file

* Menubar: Split cases of '.v' vs '.mv' file

* Update issue templates (#69)

* Update issue templates

* Update issue templates

* Styling: Unify the styling of coq code and coq doc

* Initial commit

* Move mappings to common location

* Choose correct `TextDocMapping` based on the file format.

* Remove debug `console.log` statements and unused imports.

* Add .vscode/settings file.

This settings gets rid of the errors telling us that `matchAll` is not supported by
telling vscode to use the correct version of typescript that we use.

* Fix no input area bug.

Previously a document without any input areas would throw
a 'failed to create pm webview' error. This has been fixed by implementing
an early return in the case that there are no input areas in the document.

Also includes nicer debug print of 'failed to create pm editor'.

* Fix import in unit test for text doc mapping.

Plus remove console.log cluttering the output.

* Tests: Update tests to reflect what the old parser did

* Fix: Coqdoc comment content should *not* include the space before the closing `*)`

* Add support for inputs and hints in v files.

Input can be created with `(* begin input *)` and `(* end input *)`.
Hint can be created with `(* begin hint : [title] *)` and `(* end hint *)`.

* Update client options to include .v file support in lsp-client.

* Only show `coq with coqdoc` tag in .mv files.

* Remove console.log from test

* Update tests for new parser

* Remove comparison files from repository.

* Remove debug diagram from repository

* Prosedoc Construction: Match old parser more exactly.

Currently only 'faults' are the 'postPostWhite', etc. tags.

* Update tests

* Translate 'verbatim' and 'quoted coq' to markdown equivalents.

* Add support for 'preformatted vernacular'.

Also fixes the regex for lists.

* Disable 'code' in coqdoc markdown.

This disables the creation of a code block in markdown using the four
spaces of identation. This does not work nicely in the coqdoc setting.

The markdown parser is now different for the coqdoc and markdown case.

* Fix newLine error and pass information to ProseMirror constructors.

The hint regex consumed newline characters and hence the newline information inside
of hint cells was incorrect. This has been fixed by changing the regex.

The `prePostWhite` and `postPreWhite` information was correctly parsed but not
passed to the ProseMirror constructor for the coqblock node, this has now been added.

* Fix newLine error and pass information to ProseMirror constructors.

The hint regex consumed newline characters and hence the newline information inside
of hint cells was incorrect. This has been fixed by changing the regex.

The `prePostWhite` and `postPreWhite` information was correctly parsed but not
passed to the ProseMirror constructor for the coqblock node, this has now been added.

* Add additional tests for coqdoc translation

* Update tests to reflect new document construction internals.

* Remove fix for "<" bug.

This fix is no longer needed with the introduction of the new way
of constructing ProseMirror documents.

* Disable editor in `.v` file case (for now)

---------

Co-authored-by: XyntaxCS <82953362+XyntaxCS@users.noreply.github.com>
Co-authored-by: jim-portegies <36723906+jim-portegies@users.noreply.github.com>