Implemented external tokenizer
Got external tokenizer semi-working for "Take ..." tactics. It currently
captures the token corresponding to the first input of the tactic, so in
"Take a,b : R", it captures "a,b".
An error still appears (can be seen in vscode's console accessible
through ctrl+shift+i) due to some out-of-bounds problem. The check that
pos < input.end should be enough to avoid this, but it isn't for some
reason.
If this approach works for the "Take..." tactic, it should hopefully be
useable for other tactics with 2+ inputs.
The second input of the tactic, meaning "R" in "Take a,b : R.", should
be captured correctly using the grammar directly in the same way that we
capture, e.g., "..."  in "We conclude that (...)."