Fix unit tests.
The previous commit caused unit tests to fail due to the way we
used the enums. Since this part of the code is no longer in use (it was
part of the old document constructor) I have simply removed it. If we
have a desire to keep it, the fix is to simply change the way the enum is used.

Also the unit test between document constructors relied on this code, but I think
we can safely trust the new document constructor.