Replace the default undo command, instead of creating our own.
Instead of creating a custom `waterproof.undo` and `waterproof.redo`
we overwrite the vscode commands `undo` and `redo` to fire our undo
and redo handlers.

To keep undo working as expected in other editors **even when Waterproof is opened to the side**,
we dispose of the commands as soon as the user blurs the editor.