Fix: Allow more than one waterproof editor to be active at a time.
This causes some difficulties with handling the undo/redo events,
these are now fixed by handling the undo/redo commands in the containing
class CoqEditorProvider and allowing the Waterproof editors to register
handlers for the undo and redo commands.

The commands are still disposed of when we switch editors.
**Even** when we switch from Waterproof to Waterproof editor, which
is not necessary, I have included a comment about this.