Update line number logic.
Showing line numbers in the editor is now controlled by a global setting
within vscode (`waterproof.showLineNumbersInEditor`).

The setting can be toggled using the command `Waterproof Toggle Line Numbers`.

When line numbers are **not** shown, the logic for computing the updated
line numbers is disabled. This hopefully causes an increase in performance.

As the setting is now globally applied (you can manually change it per workspace)
the line numbers show up on opening the editor when turned on.